hide delete button if minItems reached instead of disabling it, extracted base functionality of quick list, implemented quick list compact (#83 & #84)
