+++ projects/material-addons/src/lib/quick-list/quick-list-compact/quick-list-compact.component.ts
@@ -1,0 +1,17 @@
+import {
+  ChangeDetectorRef,
+  Component,
+} from '@angular/core';
+import { BaseQuickListComponent, QuickListItem } from '../base-quick-list.component';
+
+@Component({
+  selector: 'mad-quick-list-compact',
+  templateUrl: './quick-list-compact.component.html',
+  styleUrls: [],
+})
+export class QuickListCompactComponent extends BaseQuickListComponent<QuickListItem> {
+
+  constructor(public changeDetectorRef: ChangeDetectorRef) {
+    super(changeDetectorRef);
+  }
+}
