+++ projects/material-addons/src/lib/quick-list/quick-list.component.html
@@ -4,12 +4,12 @@
 <ng-container *ngFor="let item of allItems">
   <div *ngIf="itemTemplate" #row>
     <ng-container *ngTemplateOutlet="itemTemplate; context: { $implicit: item }"></ng-container>
-    <mad-icon-button *ngIf="!(readonly || readonly === '')" [disabled]="!isDeleteAllowed()" (click)="removeItem(item)">
+    <mad-icon-button *ngIf="!readonly && isDeleteAllowed()" (click)="removeItem(item)">
       <mat-icon>delete</mat-icon>
     </mad-icon-button>
   </div>
 </ng-container>
-<mad-link-button *ngIf="!(readonly || readonly === '')" (click)="addItem()"
+<mad-link-button *ngIf="!readonly" (click)="addItem()"
                  [disabled]="!addPossible || !isAddAllowed()">
   {{ addLabel }}
 </mad-link-button>
