+++ src/app/example-components/quick-list-compact-basic/quick-list-compact-basic.component.scss
@@ -1,0 +1,8 @@
+mad-readonly-form-field-wrapper:not(:first-child) {
+  margin-left: 2em;
+}
+
+.input-wrapper {
+  width: 20em;
+  margin-bottom: 2em;
+}
